requirements.txt
Django>=5.2
mysqlclient>=2.2.0
Pillow>=10.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5
drf-yasg>=1.21.7
openpyxl>=3.1.2
reportlab>=4.0.0
python-dateutil>=2.8.2

# Production server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Environment variables
python-decouple>=3.8

# Redis (optional - for caching and Celery)
redis>=5.0.0
django-redis>=5.4.0

# Monitoring (optional)
# sentry-sdk>=1.40.0
